requirements.txt
#python
# Python 3.9.6
python-dotenv==1.0.1
flask-mail==0.10.0
accelerate==0.29.3
Flask==3.0.3
annotated-types==0.7.0
anyio==4.4.0
bitsandbytes==0.42.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.21.1
openai==1.35.10
packaging==24.1
psutil==6.0.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
python-docx==0.8.11
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.10.1
sniffio==1.3.1
SpeechRecognition==3.8.1
sympy==1.12.1
tokenizers==0.19.1
torch==2.0.1
tqdm==4.66.4
transformers==4.40.0
typing_extensions==4.12.2
urllib3==2.2.2